+++ frontend/src/static/js/utils/hooks/index.ts
@@ -4,7 +4,6 @@ export * from './useItemListInlineSlider';
 export * from './useItemListLazyLoad';
 export * from './useItemListSync';
 export * from './useLayout';
-export * from './useManagementTableHeader';
 export * from './useMediaFilter';
 export * from './useMediaItem';
 export * from './usePopup';