+++ frontend/src/templates/config/core/api.config.js
@@ -7,6 +7,4 @@ module.exports = {
   members: '/users',
   history: '/user/action/watch',
   actions: '/actions',
-  manage_media: '/manage_media',
-  manage_users: '/manage_users',
 };